+++ lib/simple_managers.dart
@@ -2,5 +2,6 @@ export 'src/flag_manager.dart';
 export 'src/index_manager.dart';
 export 'src/multi_flag_manager.dart';
 export 'src/multi_index_manager.dart';
+export 'src/state_manager.dart';
 export 'src/textfield_manager.dart';
 export 'src/value_manager.dart';