+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.0.0
+
+* Added StateManager class. This class is a manager class that includes all the manager classes in this package and is easy to use.
+* For IndexManager and TextFieldManager, the keys when saving have changed, and the data structure has been changed to be similar to other manager classes. Please be careful if you are dealing with these directly in the backend.
+* Added setText and getText method to TextFieldManager.
+
 ## 2.0.0
 
 * The fromDict method has been changed to a named constructor.